Formatting adjustments again

+++ eslint.config.mjs
@@ -33,7 +33,7 @@ export default [
             "@typescript-eslint/ban-ts-comment": "off",
             "@typescript-eslint/ban-types": "off",
             "@typescript-eslint/no-require-imports": "off",
-            quotes: ["error", "double"],
+            quotes: ["error", "single"],
             semi: ["error", "always"],
             "no-loss-of-precision": "error",
             "no-template-curly-in-string": "error",
